Use selected local in LocalizationProvider
