+++ client/packages/common/src/intl/utils/DateUtils.ts
@@ -1,4 +1,4 @@
-import { SupportedLocales, useIntlUtils } from '@common/intl';
+import { getLocale, useIntlUtils } from '@common/intl';
 import {
   addMinutes,
   addDays,
@@ -33,17 +33,6 @@ import {
   previousMonday,
   endOfWeek,
 } from 'date-fns';
-// importing individually to reduce bundle size
-// the date-fns methods are tree shaking correctly
-// but the locales are not. when adding, please add as below
-import enGB from 'date-fns/locale/en-GB';
-import enUS from 'date-fns/locale/en-US';
-import fr from 'date-fns/locale/fr';
-import ar from 'date-fns/locale/ar';
-import es from 'date-fns/locale/es';
-
-// Map locale string (from i18n) to locale object (from date-fns)
-const getLocaleObj = { fr, ar, es };
 
 export const MINIMUM_EXPIRY_MONTHS = 3;
 
@@ -150,19 +139,6 @@ export const DateUtils = {
   HOUR,
   /** Number of milliseconds in one day */
   DAY,
-};
-
-const getLocale = (language: SupportedLocales) => {
-  switch (language) {
-    case 'en':
-      return navigator.language === 'en-US' ? enUS : enGB;
-    case 'tet':
-      return enGB;
-    case 'fr-DJ':
-      return fr;
-    default:
-      return getLocaleObj[language];
-  }
 };
 
 export const useFormatDateTime = () => {
